Change bottom sheet behaviour
- Refactor to a single fragment.
- Extract BottomSheet behaviour to a custom class.
