+++ android/app/src/main/res/menu/bottom_nav_menu.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
-        android:id="@+id/mapsFragment"
+        android:id="@+id/findMealFragment"
         android:icon="@drawable/ic_baseline_map_24"
         android:title="@string/map_tab_label" />
