+++ android/app/src/main/res/navigation/main_nav_graph.xml
@@ -3,22 +3,18 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main_nav_graph"
-    app:startDestination="@id/mapsFragment">
+    app:startDestination="@id/findMealFragment">
 
     <fragment
-        android:id="@+id/mapsFragment"
-        android:name="com.kinandcarta.feature.find.meal.view.MapsFragment"
+        android:id="@+id/findMealFragment"
+        android:name="com.kinandcarta.feature.find.meal.view.FindMealFragment"
         android:label="fragment_maps"
-        tools:layout="@layout/fragment_maps" />
+        tools:layout="@layout/find_meal_fragment" />
     <fragment
         android:id="@+id/settingsFragment"
         android:name="com.kcc.kmmhackathon.androidHackathonApp.view.SettingsFragment"
         android:label="settings_fragment"
         tools:layout="@layout/settings_fragment" />
-    <fragment
-        android:id="@+id/findMealFragment"
-        android:name="com.kinandcarta.feature.find.meal.view.FindMealFragment"
-        android:label="FindMealFragment" />
     <fragment
         android:id="@+id/addMealFragment"
         android:name="com.kinandcarta.lib.add.meal.view.AddMealFragment"
