+++ android/feature/find-meal/src/main/res/layout/find_meal_fragment.xml
@@ -6,6 +6,14 @@
     android:layout_height="match_parent"
     tools:context=".view.FindMealFragment">
 
+    <fragment
+        android:id="@+id/map"
+        android:name="com.google.android.gms.maps.SupportMapFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context="com.kinandcarta.feature.find.meal.view.MapsFragment">
+    </fragment>
+
     <LinearLayout
         android:id="@+id/list_bottomSheet"
         android:fitsSystemWindows="true"
@@ -23,7 +31,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
-            android:text="@string/title_switch_to_list"
+            android:text="@string/title_show_list"
             android:paddingBottom="@dimen/default_padding"
             android:textColor="@color/colorPrimary"
             android:textAppearance="@style/TextAppearance.AppCompat.Large" />
@@ -31,6 +39,7 @@
             <androidx.recyclerview.widget.RecyclerView
                 android:id="@+id/rvMeals"
                 android:layout_width="match_parent"
+                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
                 android:layout_height="match_parent" />
 
     </LinearLayout>
