--- android/feature/find-meal/src/main/res/layout/fragment_maps.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mapCoordinator"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <fragment
-        android:id="@+id/map"
-        android:name="com.google.android.gms.maps.SupportMapFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context="com.kinandcarta.feature.find.meal.view.MapsFragment">
-    </fragment>
-
-    <fragment
-        android:name="com.kinandcarta.feature.find.meal.view.FindMealFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
